공용/전용면적 data type 변경: decimal(3,2) -> decimal(6,2)

+++ modules/adapter/infrastructure/sqlalchemy/persistence/model/warehouse/right_lot_out_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, BigInteger, Integer, Float, Boolean
+from sqlalchemy import Column, String, BigInteger, Integer, Float, Boolean, Numeric
 
 from modules.adapter.infrastructure.sqlalchemy.mapper import warehouse_base
 from modules.adapter.infrastructure.sqlalchemy.persistence.model.mixins.timestamp_mixin import (
@@ -22,8 +22,8 @@ class RightLotOutModel(warehouse_base, TimestampMixin):
     deal_year = Column(String(4), nullable=True)
     deal_month = Column(String(2), nullable=True)
     deal_day = Column(String(6), nullable=True)
-    private_area = Column(Float, nullable=True)
-    supply_area = Column(Float, nullable=True)
+    private_area = Column(Numeric(6, 2), nullable=True)
+    supply_area = Column(Numeric(6, 2), nullable=True)
     regional_cd = Column(String(5), nullable=True, index=True)
     floor = Column(String(4), nullable=True)
     is_available = Column(Boolean, nullable=True)
